+++ getmac/getmac.py
@@ -388,6 +388,14 @@ def _read_arp_file(host):
     return None
 
 
+def _arping(host):
+    # type: (str) -> Optional[str]
+    return _search(
+        r" from %s \[(%s)\]" % (re.escape(host), MAC_RE_COLON),
+        _popen("arping", "-f -c 1 %s" % host),
+    )
+
+
 def _read_file(filepath):
     # type: (str) -> Optional[str]
     try:
@@ -525,6 +533,9 @@ def _hunt_for_mac(to_find, type_of_thing, net_ok=True):
             ),
             _uuid_ip,
         ]
+        # Add methods that make network requests
+        if net_ok and type_of_thing != IP6:
+            methods.append(_arping)
     else:
         log.critical("Reached end of _hunt_for_mac() if-else chain!")
         return None