+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 * Added support for Thomas Habets' version of `arping` in addition to the existing iputils one
+* Added support for docker in network bridge mode
 
 ## 0.8.2 (12/07/2019)